+++ src/main/kotlin/com/seekheart/superpal/models/web/RaidResponse.kt
@@ -2,4 +2,4 @@ package com.seekheart.superpal.models.web
 
 import java.util.*
 
-data class RaidResponse(val id: UUID, val tier: Long, var state: RaidStatusOptions, val bosses: List<Boss>)
+data class RaidResponse(val id: UUID, val tier: Int, var state: RaidStatusOptions, val bosses: List<Boss>)
